Split up dialog to use common component
