+++ src/components/components/ebay-dialog-base/index.marko
@@ -1,0 +1,34 @@
+import processHtmlAttributes from "../../../common/html-attributes"
+
+static var ignoredAttributes = ["open", "type", "focus", "a11yCloseText", "class", "windowClass"];
+
+<div
+    ...processHtmlAttributes(input, ignoredAttributes)
+    aria-modal="true"
+    role="dialog"
+    class=[input.type, input.class]
+    hidden:no-update=!state.open
+    onClick("handleDialogClick")
+    onMousedown("handleStartClick")>
+    <div key="window" class=[`${input.type}__window`, input.windowClass]>
+        <${input.header ? "div" : null} class=`${input.type}__header`>
+            <button
+                key="close"
+                class=`${input.type}__close`
+                type="button"
+                aria-label=input.a11yCloseText
+                onClick("handleCloseButtonClick")>
+                <ebay-icon name="close"/>
+            </button>
+            <if(input.header)>
+                <${input.header.renderBody} />
+            </if>
+        </>
+        <div class=`${input.type}__body` key="body">
+            <${input.renderBody}/>
+        </div>
+        <div class=`${input.type}__footer` if(input.footer)>
+            <${input.footer.renderBody} />
+        </div>
+    </div>
+</div>
