+++ src/components/ebay-drawer/examples/01-default/template.marko
@@ -1,0 +1,21 @@
+<ebay-drawer a11y-close-text="Close drawer" aria-labelledby="drawer-title" open=state.open on-drawer-close('closeDrawer')>
+    <@header><h2 id="drawer-title">Heading</h2></@header>
+    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+    <p><a href="http://www.ebay.com">www.ebay.com</a></p>
+</ebay-drawer>
+
+<ebay-button on-button-click('opendrawer')>
+    Opendrawer
+</ebay-button>
+
+class {
+  onCreate() {
+    this.state = { open: false };
+  }
+  openDrawer() {
+    this.state.open = true;
+  }
+  closeDrawer() {
+    this.state.open = false;
+  }
+}
